+++ test/jenkins.js
@@ -1,6 +1,7 @@
 var bag = require('bagofholding'),
   buster = require('buster'),
   dgram = require('dgram'),
+  feedparser = require('feedparser'),
   Jenkins = require('../lib/jenkins'),
   request = require('request');
 
@@ -628,6 +629,43 @@ buster.testCase('jenkins - version', {
     jenkins.version(function (err, result) {
       assert.isNull(err);
       assert.equals(result, '1.464');
+      done();
+    });
+  }
+});
+
+buster.testCase('jenkins - feed', {
+  setUp: function () {
+    this.mockFeedParser = this.mock(feedparser);
+  },
+  'should parse jenkins feed articles when job name is not provided': function (done) {
+    this.mockFeedParser.expects('parseUrl').once().withArgs('http://localhost:8080/rssAll').callsArgWith(1, null, null, [ { title: 'some title 1' }, { title: 'some title 2' }]);
+    var jenkins = new Jenkins('http://localhost:8080');
+    jenkins.feed(undefined, function (err, result) {
+      assert.isNull(err);
+      assert.equals(result.length, 2);
+      assert.equals(result[0].title, 'some title 1');
+      assert.equals(result[1].title, 'some title 2');
+      done();
+    });
+  },
+  'should parse job feed articles when job name is provided': function (done) {
+    this.mockFeedParser.expects('parseUrl').once().withArgs('http://localhost:8080/job/somejob/rssAll').callsArgWith(1, null, null, [ { title: 'some title 1' }, { title: 'some title 2' }]);
+    var jenkins = new Jenkins('http://localhost:8080');
+    jenkins.feed('somejob', function (err, result) {
+      assert.isNull(err);
+      assert.equals(result.length, 2);
+      assert.equals(result[0].title, 'some title 1');
+      assert.equals(result[1].title, 'some title 2');
+      done();
+    });
+  },
+  'should error to callback when an error occurs': function (done) {
+    this.mockFeedParser.expects('parseUrl').once().withArgs('http://localhost:8080/job/somejob/rssAll').callsArgWith(1, new Error('some error'));
+    var jenkins = new Jenkins('http://localhost:8080');
+    jenkins.feed('somejob', function (err, result) {
+      assert.equals(err.message, 'some error');
+      assert.equals(result, undefined);
       done();
     });
   }