+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.1.2-pre
-*
+* Move proxy environment variable handling to bag.http.requuest and bag.http.proxy
 
 ### 0.1.1
 * Move status colouring to cli so that when lib/jenkins is used programatically then it gets plain uncoloured text